When forwarding a miscellaneous exit to "error", give the best message possible.
If the exit had an Error as output, use its message.  If it had a non-Error, inspect the output and use that.  If it had no output at all, use the exit description.

Also use a pre-built Error object in `exec` to get a nicer stack trace for error messages.